Revert "update : 未使用だったイニシャライズ関数をタイトルシーンに割り当て"
This reverts commit 7953291a5e7b6ce98793242f24fc64e725c9a192.

+++ PP2nd/PP2nd/define.h
@@ -323,8 +323,8 @@ inline Vector2Int WorldPos2ArrayPos(VECTOR pos)
 {
 	Vector2Int result =
 	{
-		static_cast<int>(floor(pos.x / MAP_UNIT)),
-		static_cast<int>(floor(pos.z / MAP_UNIT))
+		floor(pos.x / MAP_UNIT),
+		floor(pos.z / MAP_UNIT)
 	};
 
 	return result;
@@ -347,7 +347,7 @@ inline string GetExecutablePath()
 /// <returns></returns>
 inline VECTOR Round(VECTOR vec, int decimal = 0)
 {
-	VECTOR result = VScale(vec, static_cast<float>(pow(10.0, decimal)));
+	VECTOR result = VScale(vec, pow(10.0, decimal));
 
 	result = VGet
 	(
@@ -356,7 +356,7 @@ inline VECTOR Round(VECTOR vec, int decimal = 0)
 		round(result.z)
 	);
 
-	result = VScale(result, static_cast<float>(pow(0.1, decimal)));
+	result = VScale(result, pow(0.1, decimal));
 
 	return result;
 }
